added 2nd sorting algorithm based on QuickSort method, which scaled better for a large number of sides

+++ src/loadbalance/loadbalance_vars.f90
@@ -78,8 +78,8 @@ REAL                                :: ElemTimePartTot  ! Total time spent for p
 REAL                                :: ElemTimeFieldTot ! Total time spent for field routines (all procs)
 REAL                                :: ElemTimePart     ! Time spent for particle routines
 REAL                                :: ElemTimeField    ! Time spent for field routines
-REAL,ALLOCATABLE                    :: ElemHDGSides(:)  ! number of master sides for the HDG solver for each element
-REAL                                :: TotalHDGSides    ! total number of master sides for the HDG solver over all local elements
+INTEGER,ALLOCATABLE                 :: ElemHDGSides(:)  ! number of master sides for the HDG solver for each element
+INTEGER                             :: TotalHDGSides    ! total number of master sides for the HDG solver over all local elements
 REAL,ALLOCATABLE                    :: ElemGlobalTime(:)
 INTEGER(KIND=8),ALLOCATABLE         :: nPartsPerElem(:)
 INTEGER(KIND=8),ALLOCATABLE         :: nDeposPerElem(:)
